gemini/backend/requirements.txt
fastapi
uvicorn
boto3
pdfplumber
numpy
faiss-cpu
sentence-transformers
requests
PyPDF2
langchain
